Refactor try two
Make all the channels part of Upstairs.
This worked, but what a mess.  A mutex for everyone.

Checking this in so I have a record, but abandoning this
as a solution.

+++ protocol/src/lib.rs
@@ -278,7 +278,7 @@ pub enum Message {
         job_id: u64,
         dependencies: Vec<u64>,
         extent_id: usize,
-        source_client_id: u8,
+        source_client_id: usize,
         source_repair_address: SocketAddr,
     },
     /// Reopen this extent, for use when upstairs is active.
